requirements.txt
# Core LangGraph & LLM
langgraph==0.0.62
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.13

# OpenAI API
openai==1.7.2

# Web Framework & UI
streamlit==1.29.0
streamlit-option-menu==0.3.6
plotly==5.17.0
pandas==2.1.4

# Data Processing
numpy==1.25.2
scikit-learn==1.3.2

# Web Scraping & APIs
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.16.0

# Database & Search
opensearch-py==2.4.2
pymongo==4.6.1

# Workflow & Scheduling
apache-airflow==2.8.0

# Configuration & Logging
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
PyYAML==6.0.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.12.1
flake8==6.1.0 